BigImage update for subsection drawing

+++ src/org/newdawn/slick/BigImage.java
@@ -286,13 +286,13 @@ public class BigImage extends Image {
 	 * @see org.newdawn.slick.Image#draw(float, float, float, float, float, float, float, float)
 	 */
 	public void draw(float x, float y, float x2, float y2, float srcx, float srcy, float srcx2, float srcy2) {
-		int srcwidth = (srcx2 - srcx)+1;
-		int srcheight = (srcy2 - srcy)+1;
+		int srcwidth = (int) (srcx2 - srcx)+1;
+		int srcheight = (int) (srcy2 - srcy)+1;
 
 		Image subImage = getSubImage(srcx,srcy,srcwidth,srcheight);
 
-		int width = (x2 - x)+1;
-		int height = (y2 - y)+1;
+		int width = (int) (x2 - x)+1;
+		int height = (int) (y2 - y)+1;
 
 		subImage.draw(x,y,width,height);
 	}
@@ -301,8 +301,8 @@ public class BigImage extends Image {
 	 * @see org.newdawn.slick.Image#draw(float, float, float, float, float, float)
 	 */
 	public void draw(float x, float y, float srcx, float srcy, float srcx2, float srcy2) {
-		int srcwidth = (srcx2 - srcx)+1;
-		int srcheight = (srcy2 - srcy)+1;
+		int srcwidth = (int) (srcx2 - srcx)+1;
+		int srcheight = (int) (srcy2 - srcy)+1;
 
 		draw(x,y,srcwidth,srcheight,srcx,srcy,srcx2,srcy2);
 	}
